Change 1: Added extensions to the error response. Change 2: Added debug option to GraphQL Configuration (flag to output exception information to extensions) Change 3: Moved serialize (), which was defined as an extension of GraphQLError, into GraphQLError. option
