+++ kgraphql/src/main/kotlin/com/apurebase/kgraphql/GraphQLError.kt
@@ -1,5 +1,6 @@
 package com.apurebase.kgraphql
 
+import com.apurebase.kgraphql.helpers.toJsonElement
 import com.apurebase.kgraphql.schema.model.ast.ASTNode
 import com.apurebase.kgraphql.schema.model.ast.Location.Companion.getLocation
 import com.apurebase.kgraphql.schema.model.ast.Source
@@ -41,12 +42,14 @@ open class GraphQLError(
      * 1. Added extensions To the error response.extensions type is Map<String, Any?>?
      * 2. Added "error type" to  make it easy to branch by error types at the client side.
     */
-    val extensionsErrorType: String? = "INTERNAL",
+    val extensionsErrorType: String? = "INTERNAL_SERVER_ERROR",
     val extensionsErrorDetail: Map<String, Any?>? = null
 ) : Exception(message) {
 
     constructor(message: String, node: ASTNode?) : this(message, nodes = node?.let(::listOf))
     constructor(message: String, extensionsErrorType: String?,extensionsErrorDetail:Map<String, Any?>?) : this(message,null,null,null,null,extensionsErrorType,extensionsErrorDetail )
+    constructor(message: String, extensionsErrorType: String?) : this(message,null,null,null,null,extensionsErrorType )
+
 
     /**
      * An array of { line, column } locations within the source GraphQL document
@@ -121,56 +124,22 @@ open class GraphQLError(
         return  exception
     }
 
-    protected fun Collection<*>.toJsonElement(): JsonElement {
-        val list: MutableList<JsonElement> = mutableListOf()
-        this.forEach {
-            val value = it as? Any ?: return@forEach
-            when(value) {
-                is Number -> list.add(JsonPrimitive(value))
-                is Boolean -> list.add(JsonPrimitive(value))
-                is String -> list.add(JsonPrimitive(value))
-                is Map<*, *> -> list.add((value).toJsonElement())
-                is Collection<*> -> list.add(value.toJsonElement())
-                is Array<*> -> list.add(value.toList().toJsonElement())
-                else -> list.add(JsonPrimitive(value.toString())) // other type
-            }
-        }
-        return JsonArray(list)
-    }
+//    protected fun buldJsonObjectByMap(it:Map<String,Any?>): JsonObject {
+//        return buildJsonObject {
+//            it.forEach { (key, value) ->
+//                when(value) {
+//                    is Number? -> put(key, value)
+//                    is String? -> put(key, value)
+//                    is Boolean? -> put(key, value)
+//                    is Map<*,*> -> put(key, value.toJsonElement())
+//                    is Collection<*> -> put(key, value.toJsonElement())
+//                    is Array<*> -> put(key, value.toList().toJsonElement())
+//                    else -> put(key, JsonPrimitive(value.toString()))  // other type
+//                }
+//            }
+//        }
+//    }
 
-    protected fun Map<*, *>.toJsonElement(): JsonElement {
-        val map: MutableMap<String, JsonElement> = mutableMapOf()
-        this.forEach {
-            val key = it.key as? String ?: return@forEach
-            val value = it.value ?: return@forEach
-            when(value) {
-                is Number? -> map[key] = JsonPrimitive(value)
-                is Boolean? -> map[key] = JsonPrimitive(value)
-                is String? -> map[key] = JsonPrimitive(value)
-                is Map<*, *> -> map[key] = (value).toJsonElement()
-                is Collection<*> -> map[key] = value.toJsonElement()
-                is Array<*> -> map[key] = value.toList().toJsonElement()
-                else -> map[key] = JsonPrimitive(value.toString())  // other type
-            }
-        }
-        return JsonObject(map)
-    }
-
-    protected fun buldJsonObjectByMap(it:Map<String,Any?>): JsonObject {
-        return buildJsonObject {
-            it.forEach { (key, value) ->
-                when(value) {
-                    is Number? -> put(key, value)
-                    is String? -> put(key, value)
-                    is Boolean? -> put(key, value)
-                    is Map<*,*> -> put(key, value.toJsonElement())
-                    is Collection<*> -> put(key, value.toJsonElement())
-                    is Array<*> -> put(key, value.toList().toJsonElement())
-                    else -> put(key, JsonPrimitive(value.toString()))  // other type
-                }
-            }
-        }
-    }
 
     open fun serialize(debug:Boolean=false): String = buildJsonObject {
         put("errors", buildJsonArray {
@@ -189,11 +158,11 @@ open class GraphQLError(
                 })
                 if (!debug) {
                     extensions?.let {
-                        put("extensions", buldJsonObjectByMap(it))
+                        put("extensions", it.toJsonElement())
                     }
                 } else {
                     extensionsDebug().let {
-                        put("extensions", buldJsonObjectByMap(it))
+                        put("extensions", it.toJsonElement())
                     }
                 }
             }
